вынесение кода, связанного с BoardTemplate
