+++ src/boardTemplate.go
@@ -1,0 +1,40 @@
+package main
+
+import _ "embed"
+
+//go:embed device_list.JSON
+var boardTemplatesRaw []byte
+
+type BoardTemplate struct {
+	ID           int      `json:"ID"`
+	VendorIDs    []string `json:"vendorIDs"`
+	ProductIDs   []string `json:"productIDs"`
+	Name         string   `json:"name"`
+	Controller   string   `json:"controller"`
+	Programmer   string   `json:"programmer"`
+	BootloaderID int      `json:"bootloaderID"`
+}
+
+// находит шаблон платы по его id
+func findTemplateByID(boardID int) *BoardTemplate {
+	var template BoardTemplate
+	if boardID < len(detector.boardTemplates) {
+		template = detector.boardTemplates[boardID]
+		// ожидается, что в файле с шаблонами прошивок (device_list.JSON) нумеровка индексов будет идти по порядку, но если это не так, то придётся перебать все шаблоны
+		if template.ID != boardID {
+			foundCorrectBootloader := false
+			for _, templ := range detector.boardTemplates {
+				if templ.ID == boardID {
+					template = templ
+					foundCorrectBootloader = true
+					break
+				}
+			}
+			if foundCorrectBootloader {
+				printLog("Не найден шаблон для устройства")
+				return nil
+			}
+		}
+	}
+	return &template
+}
